# net/http: performance regression on http.ServeFile