… src/net/http/testdata/issue61530.json