The change to io.Copy in servefile caused a exponential increase in
memory allocations if the reader is a os.File.

```
             │   sec/op    │   sec/op     vs base                │
ServeFile-16   9.165µ ± 7%   3.759µ ± 3%  -58.99% (p=0.000 n=10)

             │     B/op     │    B/op     vs base                │
ServeFile-16   33358.0 ± 0%   620.0 ± 0%  -98.14% (p=0.000 n=10)

             │ allocs/op  │ allocs/op     vs base               │
ServeFile-16   14.00 ± 0%   15.00 ± 0%    +7.14% (p=0.000 n=10)
```

Fixes: https://github.com/golang/go/issues/61530