@mauri870 2023-07-24
Closed as a solved by https://go-review.googlesource.com/c/go/+/512615